WIP: reports router
